fix(kvark): vis arrangementer i adminpanelet for gruppeledere
En gruppeleder så bare «Grupper» og «Roller og verv» i adminpanelet, og
kom dermed aldri fram til å opprette arrangementer på vegne av gruppa si
— de to eneste oppføringene som har `allowGroupLeader`. Resten av panelet
gates på rettighetsstrengene i sesjonen.

Serveren gir hver leder LEADER_BASELINE_PERMISSIONS scoped til gruppa
(69a60c1), så strengene *skal* ligge der. Klienten stolte blindt på at
API-et hadde regnet dem ut, og da er panelet bare så oppdatert som den
API-en som svarte: et image som ikke har rullet ut ennå, eller et
medlem som nettopp er gjort til leder, gir en leder som ikke får se
jobben sin. Prod kjører i skrivende stund eldre kode enn siste tag —
openapi mangler både `baseline` på GroupLeaderPermissions og
/user/:id/approve — så det er ikke et hypotetisk tilfelle.

Kvark utleder nå de samme rettighetene selv fra session.groups, i
getAuthSession, altså før både loader-context og query-cache. Lista
flyttes til en importfri modul (@photon/auth/rbac/leader-baseline) slik
at frontend kan lese den uten å dra databaselaget inn i bundlen;
checker.ts re-eksporterer den, så backend er uendret.

Fordi alle gatene leser samme liste treffer det hele kjeden på én gang:
sidebar, dashboard-snarveien, «Nytt arrangement», arrangørgruppe-
velgeren, redigering, oppmøte og betalingsoversikt. Det kan aldri vise
en leder mer enn egen gruppes arrangementer, og API-et avgjør fortsatt —
det bruker nøyaktig samme liste. Prikker og refusjon er utenfor, som før.
